Created NutrientGoalScreen

+++ core/src/main/java/com/h_fahmy/core/data/preferences/UserInfoDataStore.kt
@@ -37,7 +37,7 @@ class UserInfoDataStore(
         }
     }
 
-    override suspend fun saveWeight(weight: Int) {
+    override suspend fun saveWeight(weight: Float) {
         dataStore.updateData {
             it.copy(weight = weight)
         }
